Big refactor to use queues rather than files.
Earlier the compression tasks wrote their data to temporary files
which were then copied to the final output file. This causes double
writes which can be slow on e.g. Raspberry Pi. Now the data is written
directly to the output file.

The data is kept in memory in queues. This changes the logic slightly,
because the writer thread must now do a streaming write from the
compression threads.

The code compiles but does not actually work yet.